scripts/requirements.txt
pyyaml>=6.0.1
requests>=2.31.0
urllib3<2.0.0